figured out why vram needed 2x block RAM
The problem is that the ICE40 block RAMs aren't true
dual port RAMS: they have a single read port and a
single write port. Yosys saw that the vram was
being read from both the host and display sides,
and simply doubled the RAM use so that each side
could read from separate mirrored block RAMS.
That exceeded the available block RAMs.

I think the solution will be for the readout module
to have a way to satisfy a request to read host data.
We'll have to sneak that in around the character
and attribute fetches. Fundamentally, I guess
FPGA_VGA will need a different way of accessing the
VRAM than HW_VGA (which had true dual port RAM for
the video memory.)