rename "vramRd" signals to "readout"
This branch is an experiment to add a second "read
port" to the vram module to serve the host_interface
module. Since block RAM has only one actual read
port in hardware, the vram module will alternate
between reading from the readout address and reading
from the host address, assigning to different data
registers for each. The only downside I can think
of is that there could be a 1-cycle delay in the
data appearing in the register, which shouldn't be
a problem for either the pixel generator or the
host interface.