redesigning host interface
I think we will be able to do bidirectional I/O on
external pins, in order to communicate with the host
data bus.

Eliminating the bank register as part of the actual
display controller. Eventually I want the VRAM to be
mapped into a single 8 KB window. If the host system
needs a bank register, it can be external to the
display controller.